Move WindowNewProgram to editor
