+++ editor/window_new_program.h
